+++ src/autocomplete.js
@@ -11,6 +11,7 @@ var lang = require("./lib/lang");
 var dom = require("./lib/dom");
 var snippetManager = require("./snippets").snippetManager;
 var config = require("./config");
+var event = require("./lib/event");
 
 /**
  * @typedef BaseCompletion
@@ -97,6 +98,19 @@ class Autocomplete {
         this.stickySelectionTimer = lang.delayedCall(function() {
             this.stickySelection = true;
         }.bind(this), this.stickySelectionDelay);
+
+        this.$firstOpenTimer = lang.delayedCall(function() {
+            var initialPosition = this.completionProvider && this.completionProvider.initialPosition;
+            if (this.autoShown || (this.popup && this.popup.isOpen) || !initialPosition) return;
+
+            var completionsForEmpty = [{
+                caption: config.nls("Loading..."),
+                value: ""
+            }];
+            this.completions = new FilteredList(completionsForEmpty);
+            this.openPopup(this.editor, initialPosition.prefix, false);
+            this.popup.renderer.setStyle("ace_loading", true);
+        }.bind(this), this.stickySelectionDelay);
     }
 
     $init() {
@@ -110,6 +124,7 @@ class Autocomplete {
         this.popup.on("show", this.$onPopupShow.bind(this));
         this.popup.on("hide", this.$onHidePopup.bind(this));
         this.popup.on("select", this.$onPopupChange.bind(this));
+        event.addListener(this.popup.container, "mouseout", this.mouseOutListener.bind(this));
         this.popup.on("changeHoverMarker", this.tooltipTimer.bind(null, null));
         return this.popup;
     }
@@ -144,8 +159,15 @@ class Autocomplete {
             if (!this.inlineRenderer.show(this.editor, completion, prefix)) {
                 this.inlineRenderer.hide();
             }
-            this.$updatePopupPosition();
+
+            // If the mouse is over the tooltip, and we're changing selection on hover don't
+            // move the tooltip while hovering over the popup.
+            if (this.popup.isMouseOver && this.setSelectOnHover) { 
+                this.tooltipTimer.call(null, null);
+                return;
+            }
         }
+        this.$updatePopupPosition();
         // @ts-expect-error TODO: potential wrong arguments
         this.tooltipTimer.call(null, null);
     }
@@ -229,6 +251,8 @@ class Autocomplete {
      * @param {boolean} [keepPopupPosition]
      */
     openPopup(editor, prefix, keepPopupPosition) {
+        this.$firstOpenTimer.cancel();
+
         if (!this.popup)
             this.$init();
 
@@ -283,6 +307,8 @@ class Autocomplete {
             this.editor.off("mousedown", this.mousedownListener);
             this.editor.off("mousewheel", this.mousewheelListener);
         }
+        this.$firstOpenTimer.cancel();
+
         this.changeTimer.cancel();
         this.hideDocTooltip();
 
@@ -330,7 +356,15 @@ class Autocomplete {
     }
 
     mousewheelListener(e) {
-        this.detach();
+        if (!this.popup.isMouseOver)
+            this.detach();
+    }
+
+    mouseOutListener(e) {
+        // Check whether the popup is still open after the mouseout event,
+        // if so, attempt to move it to its desired position.
+        if (this.popup.isOpen)
+            this.$updatePopupPosition();
     }
 
    goTo(where) {
@@ -433,7 +467,10 @@ class Autocomplete {
         var prefix = util.getCompletionPrefix(this.editor);
         this.base = session.doc.createAnchor(pos.row, pos.column - prefix.length);
         this.base.$insertRight = true;
-        var completionOptions = { exactMatch: this.exactMatch };
+        var completionOptions = {
+            exactMatch: this.exactMatch,
+            ignoreCaption: this.ignoreCaption
+        };
         this.getCompletionProvider({
             prefix,
             pos
@@ -445,6 +482,7 @@ class Autocomplete {
             function (err, completions, finished) {
                 var filtered = completions.filtered;
                 var prefix = util.getCompletionPrefix(this.editor);
+            this.$firstOpenTimer.cancel();
 
                 if (finished) {
                     // No results
@@ -452,9 +490,8 @@ class Autocomplete {
                         var emptyMessage = !this.autoShown && this.emptyMessage;
                         if (typeof emptyMessage == "function") emptyMessage = this.emptyMessage(prefix);
                         if (emptyMessage) {
-                            var completionsForEmpty = [
-                                {
-                                    caption: this.emptyMessage(prefix),
+                        var completionsForEmpty = [{
+                            caption: emptyMessage,
                                     value: ""
                                 }
                             ];
@@ -475,8 +512,13 @@ class Autocomplete {
                 }
                 this.completions = completions;
                 this.openPopup(this.editor, prefix, keepPopupPosition);
-            }.bind(this)
-        );
+
+            this.popup.renderer.setStyle("ace_loading", !finished);
+        }.bind(this));
+
+        if (!this.autoShown && !(this.popup && this.popup.isOpen)) {
+            this.$firstOpenTimer.delay(this.stickySelectionDelay/2);
+        }
     }
 
     cancelContextMenu() {
@@ -485,7 +527,7 @@ class Autocomplete {
 
     updateDocTooltip() {
         var popup = this.popup;
-        var all = popup.data;
+        var all = this.completions.filtered;
         var selected = all && (all[popup.getHoveredRow()] || all[popup.getRow()]);
         var doc = null;
         if (!selected || !this.editor || !this.popup.isOpen)