+++ src/autocomplete/popup.js
@@ -111,6 +111,7 @@ class AcePopup {
             }
             lastMouseEvent = e;
             lastMouseEvent.scrollTop = popup.renderer.scrollTop;
+            popup.isMouseOver = true;
             var row = lastMouseEvent.getDocumentPosition().row;
             if (hoverMarker.start.row != row) {
                 if (!hoverMarker.id) popup.setRow(row);
@@ -164,7 +165,10 @@ class AcePopup {
             return hoverMarker.start.row;
         };
 
-        event.addListener(popup.container, "mouseout", hideHoverMarker);
+        event.addListener(popup.container, "mouseout", function() {
+            popup.isMouseOver = false;
+            hideHoverMarker();
+        });
         popup.on("hide", hideHoverMarker);
         popup.on("changeSelection", hideHoverMarker);
 
@@ -237,6 +241,7 @@ class AcePopup {
         popup.isTopdown = false;
         popup.autoSelect = true;
         popup.filterText = "";
+        popup.isMouseOver = false;
 
         popup.data = [];
         popup.setData = function (list, filterText) {
@@ -439,6 +444,7 @@ dom.importCssString(`
     margin-left: 0.9em;
 }
 .ace_completion-message {
+    margin-left: 0.9em;
     color: blue;
 }
 .ace_editor.ace_autocomplete .ace_completion-highlight{
@@ -483,6 +489,29 @@ dom.importCssString(`
 }
 .ace_autocomplete .ace_completion-spacer {
     flex: 1;
+}
+.ace_autocomplete.ace_loading:after  {
+    content: "";
+    position: absolute;
+    top: 0px;
+    height: 2px;
+    width: 8%;
+    background: blue;
+    z-index: 100;
+    animation: ace_progress 3s infinite linear;
+    animation-delay: 300ms;
+    transform: translateX(-100%) scaleX(1);
+}
+@keyframes ace_progress {
+    0% { transform: translateX(-100%) scaleX(1) }
+    50% { transform: translateX(625%) scaleX(2) } 
+    100% { transform: translateX(1500%) scaleX(3) } 
+}
+@media (prefers-reduced-motion) {
+    .ace_autocomplete.ace_loading:after {
+        transform: translateX(625%) scaleX(2);
+        animation: none;
+     }
 }
 `, "autocompletion.css", false);